requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
hyperliquid-python-sdk>=0.18.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
python-telegram-bot>=20.0
torch>=2.0.0
scikit-learn>=1.2.0
eth-account>=0.8.0
